requirements.txt
bitvec==0.2.7
click==8.1.6
clipboard==0.0.4
jsmin==3.0.1
matplotlib==3.5.2
NBT==1.5.1
numpy==1.23.5
pandas==1.5.3
path.py==12.5.0
python_Levenshtein==0.21.1
Requests==2.31.0
scikit_learn==1.2.2
scipy==1.11.2
tabulate==0.9.0